extracted selection interface back to its own file
